Extract CourseBuilder to Interface
